TheMicTestingFunction
Once clicked on the Mic button a progressbar shows up indicating microphone testing in progress. And anything that the microphone detected is returned to the speaker after a short delay indicating the functionality of the microphone.
The testing process is terminated with another click on the mic button.
Tried to import an microphone Icon but failed, don't know the reason yet.
That is propobaly all for this update.